사용자 상세 조회 완료

+++ src/main/java/com/meongnyang/shop/dto/request/admin/ReqModifyProductDto.java
@@ -26,7 +26,7 @@ public class ReqModifyProductDto {
     private Long categoryId;
     @NotNull
     private Long productPrice;
-    private Long productPriceDiscount;
+    private int productPriceDiscount;
     private String productDetail;
     private String productBrand;
     private String productModel;
